feat(error handling): refactor avatar and integrate with error handling
